requirements.txt
Flask==1.1.2
sklearn
numpy
google-cloud-storage
google-cloud-bigquery
